\name{gibby-package}
\alias{gibby-package}
\alias{gibby}
\docType{package}
\title{\packageTitle{gibby}}
\description{\packageDescription{gibby}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{gibby}
  \packageIndices{gibby}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{gibby}
  
  Maintainer: \packageMaintainer{gibby}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
